pyproject.toml
[project]
name = "statswr-mcp-server"
version = "0.1.0"
description = "An MCP server that allows AI agents to interact with the statsWR api"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
]
